requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.15.0
azure-eventgrid==4.11.0
azure-communication-callautomation==1.3.0
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0
azure-storage-blob==12.25.1

# development
azure-communication-identity